Post Views and URl mapping done,
MASSIVE push! alot of new content established into the web app.

CRUD for the post!

the majority of this push is setting up the CRUD views for the post.  many things we're learned such as ORM with CBV. method of using SQL with python to get selected item or oder items.

Nn introduction to,  login required mixins. an analogy to the login required decorator for function based views but instead the use of Class based views. as well as some functions that needs to be set up when using this such as 'login_url' that  redirect a user thats not logged into login page.

further set up to the urls.py to connect the views. as well as setting up the file structure.