base.html set up
setting up the base.html that other templates will inherit!

added all te CDN the website will use ranging from bootstrap, to fonts.

we created the nav bar contain links to the main pages for this website - logging in and out, create post and drafting a post.

little CSS was done in this push