about us & home page/post_list page done!
this push was simply setting up all the temaplates/html pages for the website for editing.

the main  thing done here is the post_list or the home page.  we created code that allows the users to see all the blog post made being able to see the list it, containing their title, published date and comments per blog item in the list.