Post_detail done!
we have now created the detail page, the page where the user view a blog post, if logged on can look at the comment.

if you are a super user you are able to make a comments, delete and edit comments. as well as edit existing blog post and delete blog post.