finshing off the templates
many new things done herre:

most of the templates done!

post and comment form templates are done. gives the user the ability to edit as well as use the thrid party software to enable format editing with their post.

post draft,  show all the post in the post, see the dated created but not posted return their title to link them to their post.

confirm deletel template, to ensure the user have a second chance to delete the form.

lastly some CSS to make it look gooood.